added map manipulation algorithms and tests (#30)
* map manip tests

* map manipulator implemented and tested

* random generator working

+++ include/color.h
@@ -25,8 +25,8 @@ namespace pathfinder {
 
         private:
         int maxAbsHeight;
-        int positiveSlope;
-        int negativeSlope;
+        double positiveSlope;
+        double negativeSlope;
     };
 
 }
